integration tets, module tests, migration

+++ src/main/java/com/start/model/dto/UserDto.java
@@ -10,6 +10,6 @@ public class UserDto {
     private String username;
     private String password;
     private int age;
-    private int salary;
+    private float salary;
 
 }
